extract lxmd three node harness
